move range-specific code out of facettermslist
