+++ src/components/browse/components/FacetList/FacetTermsList.js
@@ -89,7 +89,7 @@ export function mergeTerms(facet, filters){
     return terms.concat(unseenTerms);
 }
 
-function segmentComponentsByStatus(termComponents){
+export function segmentComponentsByStatus(termComponents){
     const groups = {};
     termComponents.forEach(function(t){
         const { props: { status } } = t;
@@ -99,14 +99,6 @@ function segmentComponentsByStatus(termComponents){
         groups[status].push(t);
     });
     return groups;
-}
-
-function getRangeStatus(range, toVal, fromVal) {
-    const { from = null, to = null } = range || {};
-    if (to === toVal && from === fromVal) {
-        return "selected";
-    }
-    return "none";
 }
 
 
@@ -415,179 +407,3 @@ export const CountIndicator = React.memo(function CountIndicator({ count = 1, co
         </svg>
     );
 });
-
-
-
-export const ListOfRanges = React.memo(function ListOfRanges(props){
-    const { facet, facetOpen, facetClosing, persistentCount = 10, onTermClick, expanded, onToggleExpanded, termTransformFxn, toVal, fromVal, resetAll } = props;
-    const { ranges = [] } = facet;
-
-    /** Create range components and sort by status (selected->omitted->unselected) */
-    const { termComponents, activeTermComponents, unselectedTermComponents,
-        totalLen, selectedLen, omittedLen, unselectedLen,
-        persistentTerms = null,
-        collapsibleTerms = null,
-        collapsibleTermsCount = 0,
-        collapsibleTermsItemCount = 0
-    } = useMemo(function(){
-        const {
-            selected: selectedTermComponents    = [],
-            omitted : omittedTermComponents     = [],
-            none    : unselectedTermComponents  = []
-        } = segmentComponentsByStatus(ranges.map(function(range){
-            return <RangeTerm {...{ facet, range, termTransformFxn, resetAll }} onClick={onTermClick} key={`${range.to}-${range.from}`} status={getRangeStatus(range, toVal, fromVal)} />;
-        }));
-
-        const selectedLen = selectedTermComponents.length;
-        const omittedLen = omittedTermComponents.length;
-        const unselectedLen = unselectedTermComponents.length;
-        const totalLen = selectedLen + omittedLen + unselectedLen;
-        const termComponents = selectedTermComponents.concat(omittedTermComponents).concat(unselectedTermComponents);
-        const activeTermComponents = termComponents.slice(0, selectedLen + omittedLen);
-
-        const retObj = { termComponents, activeTermComponents, unselectedTermComponents, selectedLen, omittedLen, unselectedLen, totalLen };
-
-        if (totalLen <= Math.max(persistentCount, selectedLen + omittedLen)) {
-            return retObj;
-        }
-
-        const unselectedStartIdx = selectedLen + omittedLen;
-        retObj.persistentTerms = []; //termComponents.slice(0, unselectedStartIdx);
-
-        var i;
-        for (i = unselectedStartIdx; i < persistentCount; i++){
-            retObj.persistentTerms.push(termComponents[i]);
-        }
-
-        retObj.collapsibleTerms = termComponents.slice(i);
-        retObj.collapsibleTermsCount = totalLen - i;
-        retObj.collapsibleTermsItemCount = retObj.collapsibleTerms.reduce(function(m, termComponent){
-            return m + (termComponent.props.range.doc_count || 0);
-        }, 0);
-
-        return retObj;
-
-    }, [ ranges, persistentCount, toVal, fromVal ]);
-
-
-
-    const commonProps = {
-        "data-any-active" : !!(selectedLen || omittedLen),
-        "data-all-active" : totalLen === (selectedLen + omittedLen),
-        "data-open" : facetOpen,
-        "className" : "facet-list",
-        "key" : "facetlist"
-    };
-
-    if (Array.isArray(collapsibleTerms) && collapsibleTerms.length > 0){
-
-        let expandButtonTitle;
-
-        if (expanded){
-            expandButtonTitle = (
-                <span>
-                    <i className="icon icon-fw icon-minus fas"/> Collapse
-                </span>
-            );
-        } else {
-            expandButtonTitle = (
-                <span>
-                    <i className="icon icon-fw icon-plus fas"/> View { collapsibleTermsCount } More
-                    <span className="pull-right">{ collapsibleTermsItemCount }</span>
-                </span>
-            );
-        }
-
-        return (
-            <div {...commonProps}>
-                <PartialList className="mb-0 active-terms-pl" open={facetOpen} persistent={activeTermComponents} collapsible={
-                    <React.Fragment>
-                        <PartialList className="mb-0" open={expanded} persistent={persistentTerms} collapsible={collapsibleTerms} />
-                        <div className="pt-08 pb-0">
-                            <div className="view-more-button" onClick={onToggleExpanded}>{ expandButtonTitle }</div>
-                        </div>
-                    </React.Fragment>
-                } />
-            </div>
-        );
-    } else {
-        return (
-            <div {...commonProps}>
-                <PartialList className="mb-0 active-terms-pl" open={facetOpen} persistent={activeTermComponents} collapsible={unselectedTermComponents} />
-            </div>
-        );
-    }
-});
-
-/**
- * Used to render a term with range functionality in FacetList. Basically same as FacetTermsList > Term... maybe merge later
- */
-export class RangeTerm extends React.PureComponent {
-
-    constructor(props){
-        super(props);
-        this.handleClick = _.debounce(this.handleClick.bind(this), 500, true);
-        this.state = {
-            'filtering' : false
-        };
-    }
-
-    handleClick(e) {
-        var { range, onClick } = this.props;
-        var { to = null, from = null } = range;
-        e.preventDefault();
-        this.setState({ 'filtering' : true }, () => {
-            onClick(to, from, e, () => this.setState({ 'filtering' : false }));
-        });
-    }
-
-    render() {
-        const { range, facet, status, termTransformFxn, resetAll } = this.props;
-        const { doc_count, from, to, label } = range;
-        const { filtering } = this.state;
-        const selected = (status !== 'none');
-        let icon = null;
-
-        let title = (
-            (typeof from !== 'undefined' ? from : '< ') +
-            (typeof from !== 'undefined' && typeof to !== 'undefined' ? ' - ' : '') +
-            (typeof to !== 'undefined' ? to : '+ ')
-        );
-
-        if (filtering) {
-            icon = <i className="icon fas icon-circle-notch icon-spin icon-fw" />;
-        } else if (status === 'selected' || status === 'omitted') {
-            icon = <i className="icon icon-dot-circle icon-fw fas" />;
-        } else {
-            icon = <i className="icon icon-circle icon-fw unselected far" />;
-        }
-
-        if (!title || title === 'null' || title === 'undefined'){
-            title = 'None';
-        }
-
-        return (
-            <li className={"facet-list-element "} key={label} data-key={label}>
-                <a className="term" data-selected={selected} href="#" onClick={status === "selected" ? resetAll : this.handleClick} data-term={label}>
-                    <span className="facet-selector">{icon}</span>
-                    <span className="facet-item" data-tip={title.length > 30 ? title : null}>{title} {label ? `(${label})` : null}</span>
-                    <span className="facet-count">{doc_count || 0}</span>
-                </a>
-            </li>
-        );
-    }
-
-}
-RangeTerm.propTypes = {
-    'facet'             : PropTypes.shape({
-        'field'             : PropTypes.string.isRequired
-    }).isRequired,
-    'range'              : PropTypes.shape({
-        'from'              : PropTypes.number,
-        'to'                : PropTypes.number,
-        'label'             : PropTypes.string,
-        'doc_count'         : PropTypes.number
-    }).isRequired,
-    // 'getTermStatus'     : PropTypes.func.isRequired,
-    'onClick'           : PropTypes.func.isRequired
-};
